Added adjust credit limit logic

+++ apps/credit-api/src/modules/margin/margin_notifications.entity.ts
@@ -20,7 +20,7 @@ export class MarginNotification {
   @Column('varchar')
   userId: string;
 
-  @Column({ type: 'int', nullable: true })
+  @Column({ type: 'float', nullable: true })
   sentAtLtv: number | null;
 
   @Column({ type: 'boolean' })
